Basic implementation of mustache to generate the Proxy and JsonRpcClient classes.
