+++ csharp/templates/Proxy.mustache
@@ -1,0 +1,100 @@
+/*
+ * Copyright (c) Citrix Systems, Inc.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   1) Redistributions of source code must retain the above copyright
+ *      notice, this list of conditions and the following disclaimer.
+ *
+ *   2) Redistributions in binary form must reproduce the above
+ *      copyright notice, this list of conditions and the following
+ *      disclaimer in the documentation and/or other materials
+ *      provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+ * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
+ * OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using CookComputing.XmlRpc;
+
+
+namespace XenAPI
+{
+    public interface Proxy : IXmlRpcProxy
+    {
+        [XmlRpcMethod("event.get_record")]
+        Response<Proxy_Event>
+        event_get_record(string session, string _event);
+
+        [XmlRpcMethod("event.get_by_uuid")]
+        Response<string>
+        event_get_by_uuid(string session, string _uuid);
+
+        [XmlRpcMethod("event.get_id")]
+        Response<string>
+        event_get_id(string session, string _event);
+
+        [XmlRpcMethod("event.set_id")]
+        Response<string>
+        event_set_id(string session, string _event, string _id);
+
+        [XmlRpcMethod("event.register")]
+        Response<string>
+        event_register(string session, string [] _classes);
+
+        [XmlRpcMethod("event.unregister")]
+        Response<string>
+        event_unregister(string session, string [] _classes);
+
+        [XmlRpcMethod("event.next")]
+        Response<Proxy_Event[]>
+        event_next(string session);
+
+        [XmlRpcMethod("event.from")]
+        Response<Events>
+        event_from(string session, string [] _classes, string _token, double _timeout);
+{{#proxy_methods}}
+{{{proxy_method}}}
+{{/proxy_methods}}
+    }
+
+    [XmlRpcMissingMapping(MappingAction.Ignore)]
+    public class Proxy_Event
+    {
+        public string id;
+        public string timestamp;
+        [XmlRpcMember("class")]
+        public string class_;
+        public string operation;
+        [XmlRpcMember("ref")]
+        public string opaqueRef;
+        [XmlRpcMember("snapshot")]
+        public object snapshot;
+    }
+{{#proxy_classes}}
+
+    [XmlRpcMissingMapping(MappingAction.Ignore)]
+    public class Proxy_{{proxy_class_name}}
+    {
+{{#proxy_class_fields}}
+        public {{proxy_field_type}} {{proxy_field_name}};
+{{/proxy_class_fields}}
+    }
+{{/proxy_classes}}
+}
